Remove unneded functionality from contract, update contracts and xsuite.
