+++ auth/meta/Cargo.toml
@@ -11,4 +11,4 @@ authors = [ "you",]
 path = ".."
 
 [dependencies.multiversx-sc-meta]
-version = "0.42.0"
+version = "0.43.4"
