+++ auth/wasm/Cargo.toml
@@ -24,4 +24,4 @@ panic = "abort"
 path = ".."
 
 [dependencies.multiversx-sc-wasm-adapter]
-version = "0.42.0"
+version = "0.43.4"
