+++ auth/wasm/src/lib.rs
@@ -10,6 +10,7 @@
 // Total number of exported functions:   7
 
 #![no_std]
+#![allow(internal_features)]
 #![feature(lang_items)]
 
 multiversx_sc_wasm_adapter::allocator!();
