+++ gateway/Cargo.toml
@@ -9,11 +9,10 @@ publish = false
 path = "src/lib.rs"
 
 [dependencies.multiversx-sc]
-version = "0.42.0"
-#features = ["promises"]
+version = "0.43.4"
 
 [dev-dependencies.multiversx-sc-scenario]
-version = "0.42.0"
+version = "0.43.4"
 
 [dev-dependencies]
 num-bigint = "0.4.2"
