+++ gateway/wasm/src/lib.rs
@@ -5,11 +5,12 @@
 ////////////////////////////////////////////////////
 
 // Init:                                 1
-// Endpoints:                           15
-// Async Callback:                       1
-// Total number of exported functions:  17
+// Endpoints:                            6
+// Async Callback (empty):               1
+// Total number of exported functions:   8
 
 #![no_std]
+#![allow(internal_features)]
 #![feature(lang_items)]
 
 multiversx_sc_wasm_adapter::allocator!();
@@ -19,22 +20,13 @@ multiversx_sc_wasm_adapter::endpoints! {
     gateway
     (
         init => init
-        sendToken => send_token
         callContract => call_contract
-        callContractWithToken => call_contract_with_token
         validateContractCall => validate_contract_call
-        validateContractCallAndMint => validate_contract_call_and_mint
         execute => execute
-        isContractCallApproved => is_contract_call_approved
-        isContractCallAndMintApproved => is_contract_call_and_mint_approved
         isCommandExecuted => is_command_executed
-        tokenMintAmount => get_token_mint_amount
-        getSupportedTokens => supported_tokens
-        transferMintLimiter => transfer_mint_limiter
-        setTokenMintLimits => set_token_mint_limits
-        mintLimiter => mint_limiter
+        isContractCallApproved => is_contract_call_approved
         authModule => auth_module
     )
 }
 
-multiversx_sc_wasm_adapter::async_callback! { gateway }
+multiversx_sc_wasm_adapter::async_callback_empty! {}
