+++ package.json
@@ -20,7 +20,7 @@
     "tsx": "^3.12.7",
     "typescript": "^5.0.4",
     "vitest": "^0.33.0",
-    "xsuite": "0.0.29",
+    "xsuite": "0.0.32",
     "keccak": "^3.0.3"
   }
 }
