+++ tests/helpers.ts
@@ -2,12 +2,10 @@ import fs from "fs"
 import { UserSecretKey } from "@multiversx/sdk-wallet/out"
 import { Address } from "@multiversx/sdk-core/out"
 import createKeccakHash from "keccak";
-import { e } from 'xsuite/data';
+import { e } from 'xsuite';
 
 export const MOCK_CONTRACT_ADDRESS_1: string = "erd1qqqqqqqqqqqqqpgqd77fnev2sthnczp2lnfx0y5jdycynjfhzzgq6p3rax";
 export const MOCK_CONTRACT_ADDRESS_2: string = "erd1qqqqqqqqqqqqqpgq7ykazrzd905zvnlr88dpfw06677lxe9w0n4suz00uh";
-
-export const DEFAULT_ESDT_ISSUE_COST = 5000000000000000;
 
 export const ALICE_ADDR = 'erd1qyu5wthldzr8wx5c9ucg8kjagg0jfs53s8nr3zpz3hypefsdd8ssycr6th';
 export const BOB_ADDR = 'erd1spyavw0956vq68xj8y4tenjpq2wd5a9p2c6j8gsz7ztyrnpxrruqzu66jx';
